% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFun.R
\name{CalcAccess}
\alias{CalcAccess}
\title{Apply NARCIS bussiness-rules for determining access-status}
\usage{
CalcAccess(rec, version = 1)
}
\arguments{
\item{rec}{A list with record-information for one record}

\item{version}{Implemented for forward compatibilty: if the business-rules change this function will be updated, but only for higher version-numbers.
Currently only 1 is implemented}
}
\value{
A nested list with information
}
\description{
To generate NARCIS-content, publications are harvested over OAI-PMH, with different schemas.
This function outputs access-status based on content.
}
