% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFun.R
\name{is.DOI}
\alias{is.DOI}
\alias{DOIify}
\title{Does a string contain a DOI, and extract it}
\usage{
is.DOI(ID)

DOIify(ID)
}
\arguments{
\item{ID}{Character vector of IDs to process}
}
\value{
for is.DOI: a logical of length(ID)
for DOIify: a character vector of length(ID) with the bare DOIs (starting with 10.). If ID contains non-DOIs they are returned unaltered
}
\description{
Basically just regular expressions, wrapped in a function to simplify adaptations
DOIify also converts to lowercase to ease comparisons. If an ID that is not a DOI is passed to DOIify it is returned unaltered.
An ID is seen as a DOI when it satisfies the regular expression "10.\.[0-9]{4,5}/"
}
