% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFun.R
\name{oa_color}
\alias{oa_color}
\title{Decide color based on Unpaywall-data}
\usage{
oa_color(..., df = NULL)
}
\arguments{
\item{..., }{df Either single vectors, or a data.frame with appropriate columns (or a list with equal-length elements)
Should either be named with the names "c('is_oa', 'host_type', 'journal_is_oa','license')", or be of length 4, in this order}
}
\value{
A vector of colors (as a factor)
}
\description{
Based on 4 criteria from Unpaywall:
\itemize{
  \item is_oa (logical)
  \item host_type (character or factor, must be 'publisher', 'repository' if is_oa is TRUE)
  \item journal_is_oa (logical)
  \item license (character or factor)
}
strings are case-insensitive
Applied rules are:
\enumerate{
  \item if is_oa is NA, 'unknown' is returned
  \item if is_oa is FALSE, 'closed' is returned
  \item if host_type is 'repository', 'green' is returned
  \item if host_type is 'publisher' and 'journal_is_oa' is TRUE, 'pure gold' is returned
  \item if host_type is 'publisher' and 'license' is NA or an empty string, 'bronze' is returned
  \item if host_type is 'publisher' and 'license' is not NA or an empty, 'hyrbid' is returned
  \item if we reach this step is_oa is TRUE, but host_type is unknown. SO we return 'unclear (error)'
}
}
