% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFun.R
\name{readNARCIScla}
\alias{readNARCIScla}
\title{read NARCIS classification scheme from pdf-file}
\usage{

  readNARCIScla(FilePath = "https://www.narcis.nl/content/pdf/classification_en.pdf")
}
\arguments{
\item{FilePath}{Path to pdf-file containing classification. May be a URL}
}
\value{
A data.frame containing codes, labels, code of overlaying level, and indication of level-depth
}
\description{
Read and convert NARCIS-classification-scheme
}
