% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFun.R
\name{ReadVSNUpdf}
\alias{ReadVSNUpdf}
\title{Read the numbers from the VSNU, based on pdf}
\usage{
ReadVSNUpdf(pdfs = paste0(getOption("Paths")$Input,
  c("/Narcis versus KUOZ 1.pdf", "/Narcis versus KUOZ 2 uitgebreid.pdf")))
}
\arguments{
\item{pdfs}{Vector with paths to the pdfs}
}
\value{
list with data.frames
}
\description{
For private use, as the underlying document is not public
}
