% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFun.R
\name{ReadVSNUxls}
\alias{ReadVSNUxls}
\title{Read the numbers from the VSNU, based downloadable Excelsheet}
\usage{

  ReadVSNUxls(xls = "http://www.vsnu.nl/files/documenten/Feiten_en_Cijfers/Inzet_en_output_2015.xlsx")
}
\arguments{
\item{xls}{Paths to excelsheet}
}
\value{
data.frame with data
}
\description{
Mostly custom, will work on only specific sheet
}
